fix: Make using portal be opt in instead
These components traps focus, while also applying aria-hidden on sibling tags and their children.

These portals make it so that the contents of their respective components goes in the <body> tag, which is bad.

For example, <Select> inside <Dialog>: when <Select> opens, it puts aria-hidden in the <Dialog>, which triggers warnings from the browser because aria-hidden should not be put in a component where a child still has focus (<Select> in this example).